feat: add Promise based API to wrap node's inspector API
Adds a class 'Profiler' to act as a promise based wrapper API to node's
inspector.Session API. This greatly simplifies the act of sending
messages to v8's inspector by removing the requirement for callback
functions. However, it introduces the potential for types to get out of
sync as a result of DefinitelyTyped relying on overloads rather than
type unions. The underlying session is exposed from this class though,
primarily with the intention of aiding debugging by allowing
[inspector.Session.prototype.post] to be called directly.

Additionally, Jest is added and configured to work with TypeScript and
Eslint rules are adjusted to play well with re-exported types and Jest
spec files.